\name{snpSOMNiBUS-package}
\alias{snpSOMNiBUS-package}
\alias{snpSOMNiBUS}
\docType{package}
\title{\packageTitle{snpSOMNiBUS}}
\description{\packageDescription{snpSOMNiBUS}}
\section{Package Content}{\packageIndices{snpSOMNiBUS}}
\author{\packageAuthor{snpSOMNiBUS}}
\section{Maintainer}{\packageMaintainer{snpSOMNiBUS}}
\keyword{package}
