\name{sparseSOMNiBUS-package}
\alias{sparseSOMNiBUS-package}
\alias{sparseSOMNiBUS}
\docType{package}
\title{\packageTitle{sparseSOMNiBUS}}
\description{\packageDescription{sparseSOMNiBUS}}
\section{Package Content}{\packageIndices{sparseSOMNiBUS}}
\author{\packageAuthor{sparseSOMNiBUS}}
\section{Maintainer}{\packageMaintainer{sparseSOMNiBUS}}
\keyword{package}
